Change ResourceManager and ShaderManager to use the new ICrFile interface, and deleted obsolete functions
Also changed the source member of ShaderStageCreate to be a CrString instead of a CrVector, as it takes care of the null terminator naturally

+++ Source/Rendering/CrImage.cpp
@@ -31,7 +31,7 @@ const void* CrImage::GetData()
 	return m_dataPointer;
 }
 
-uint32_t CrImage::GetDataSize()
+uint64_t CrImage::GetDataSize()
 {
 	return m_dataSize;
 }
